[REFACTOR] : ♻️ 프로필 이미지 클릭 시, 화면이동 함수 모듈화
