+++ IDo/Page/NoticeBoard/NoticeBoardViewController.swift
@@ -142,43 +142,8 @@ extension NoticeBoardViewController: UITableViewDelegate, UITableViewDataSource 
     }
     func navigateToProfilePage(for indexPath: IndexPath) {
         let profile = firebaseManager.noticeBoards[indexPath.row].rootUser
-        let profileViewController = MyProfileViewController()
-        profileViewController.userProfile = profile
-
-        if let profileImageURL = profile.profileImagePath {
-            firebaseManager.getUserImage(referencePath: profileImageURL, imageSize: .medium) { [weak profileViewController] downloadedImage in
-                DispatchQueue.main.async {
-                    if let image = downloadedImage {
-                        profileViewController?.profileImage.setImage(image, for: .normal)
-                    }
-                }
-            }
-        }
-        else {
-            
-            if let defaultImage = UIImage(named: "profile") {
-                profileViewController.profileImage.setImage(defaultImage, for: .normal)
-            }
-        }
-
-        profileViewController.profileName.text = profile.nickName
-        if let hobbyList = profile.hobbyList {
-            profileViewController.choiceEnjoyTextField.text = hobbyList.first
-        }
-        profileViewController.selfInfoDetail.text = profile.description
-        
-        profileViewController.profileImage.isUserInteractionEnabled = true
-        profileViewController.profileName.isEditable = false
-        profileViewController.choicePickerView.isUserInteractionEnabled = false
-        profileViewController.selfInfoDetail.isEditable = false
-        profileViewController.logout.isHidden = true
-        profileViewController.line.isHidden = true
-        profileViewController.deleteID.isHidden = true
-
-         // 전체 화면으로 모달을 표시하려면 이 줄을 추가하세요.
-        self.present(profileViewController, animated: true, completion: nil)
+        PresentToProfileVC.presentToProfileVC(from: self, with: profile)
     }
-
 }
 
 // MARK: - FirebaseManaerDelegate 관련
