+++ IDo/Page/NoticeBoard/NoticeHomeController.swift
@@ -319,60 +319,8 @@ extension NoticeHomeController: UITableViewDelegate, UITableViewDataSource {
     
     func navigateToProfilePage(for indexPath: IndexPath) {
         if let userList = firebaseClubDatabaseManager.model?.userList {
-            let profileViewController = MyProfileViewController()
             let profile = userList[indexPath.row]
-            profileViewController.userProfile = profile
-
-            if let profileImageURL = profile.profileImagePath {
-//                FBURLCache.shared.downloadURL(storagePath: profileImageURL + "/\(ImageSize.small.rawValue)") { result in
-//                    switch result {
-//                    case .success(let image):
-//                        DispatchQueue.main.async {
-//                            profileViewController.profileImage.setImage(image, for: .normal)
-//                        }
-//                    case .failure(let error):
-//                        print(error.localizedDescription)
-//                    }
-//                }
-                getUserImage(referencePath: profileImageURL, imageSize: .medium) { [weak profileViewController] downloadedImage in
-                    DispatchQueue.main.async {
-                        if let image = downloadedImage {
-                            profileViewController?.profileImage.setImage(image, for: .normal)
-                        }
-                    }
-                }
-            }
-            else {
-                let defaultImage = UIImage(named: "profile") ?? UIImage(systemName: "person.fill")
-                profileViewController.profileImage.setImage(defaultImage, for: .normal)
-            }
-
-            profileViewController.profileName.text = profile.nickName
-            profileViewController.choiceEnjoyTextField.text = profile.hobbyList?.first
-            profileViewController.selfInfoDetail.text = profile.description
-            
-            profileViewController.profileImage.isUserInteractionEnabled = true
-            profileViewController.profileName.isEditable = false
-            profileViewController.choicePickerView.isUserInteractionEnabled = false
-            profileViewController.selfInfoDetail.isEditable = false
-            profileViewController.logout.isHidden = true
-            profileViewController.line.isHidden = true
-            profileViewController.deleteID.isHidden = true
-            
-            self.present(profileViewController, animated: true, completion: nil)
-        }
-    }
-    
-    func getUserImage(referencePath: String?, imageSize: ImageSize, completion: @escaping(UIImage?) -> Void) {
-        guard let referencePath else { return }
-        let imageRefPath = storage.child(referencePath).child(imageSize.rawValue).fullPath
-        self.urlCache.downloadURL(storagePath: imageRefPath) { result in
-            switch result {
-            case .success(let image):
-                completion(image)
-            case .failure(let error):
-                print(error.localizedDescription)
-            }
+            PresentToProfileVC.presentToProfileVC(from: self, with: profile)
         }
     }
 }
