+++ IDo/Page/NoticeBoardDetailPage/NoticeBoardDetailViewController.swift
@@ -177,39 +177,7 @@ private extension NoticeBoardDetailViewController {
     
     func presentToProfilePage() {
         let profile = noticeBoard.rootUser
-        let profileViewController = MyProfileViewController()
-        profileViewController.userProfile = profile
-        
-        if let profileImageURL = profile.profileImagePath {
-            getUserImage(referencePath: profileImageURL, imageSize: .medium) { [weak profileViewController] downloadedImage in
-                DispatchQueue.main.async {
-                    if let image = downloadedImage {
-                        profileViewController?.profileImage.setImage(image, for: .normal)
-                    }
-                }
-            }
-        }
-        else {
-            
-            if let defaultImage = UIImage(named: "profile") {
-                profileViewController.profileImage.setImage(defaultImage, for: .normal)
-            }
-        }
-        
-        profileViewController.profileName.text = profile.nickName
-        profileViewController.choiceEnjoyTextField.text = profile.hobbyList?.first
-        profileViewController.selfInfoDetail.text = profile.description
-        
-        profileViewController.profileImage.isUserInteractionEnabled = true
-        profileViewController.profileName.isEditable = false
-        profileViewController.choicePickerView.isUserInteractionEnabled = false
-        profileViewController.selfInfoDetail.isEditable = false
-        profileViewController.logout.isHidden = true
-        profileViewController.line.isHidden = true
-        profileViewController.deleteID.isHidden = true
-        
-        self.present(profileViewController, animated: true, completion: nil)
-    
+        PresentToProfileVC.presentToProfileVC(from: self, with: profile)
     }
     
     func noticeBoardSetup() {
@@ -628,52 +596,6 @@ extension NoticeBoardDetailViewController: UITableViewDelegate, UITableViewDataS
     }
     func navigateToProfilePage(for indexPath: IndexPath) {
         let profile = firebaseCommentManager.modelList[indexPath.row].writeUser
-        let profileViewController = MyProfileViewController()
-        profileViewController.userProfile = profile
-
-        // If the profile image path is not nil, fetch and set the image
-        if let profileImageURL = profile.profileImagePath {
-            firebaseCommentManager.getUserImage(referencePath: profileImageURL, imageSize: .medium) { [weak profileViewController] downloadedImage in
-                DispatchQueue.main.async {
-                    if let image = downloadedImage {
-                        profileViewController?.profileImage.setImage(image, for: .normal)
-                    }
-                }
-            }
-        } else {
-            // Set a default image if there is no profileImagePath
-            if let defaultImage = UIImage(named: "profile") {
-                profileViewController.profileImage.setImage(defaultImage, for: .normal)
-            }
-        }
-
-        profileViewController.profileName.text = profile.nickName
-        profileViewController.choiceEnjoyTextField.text = profile.hobbyList?.first
-        profileViewController.selfInfoDetail.text = profile.description
-        
-        profileViewController.profileImage.isUserInteractionEnabled = true
-        profileViewController.profileName.isEditable = false
-        profileViewController.choicePickerView.isUserInteractionEnabled = false
-        profileViewController.selfInfoDetail.isEditable = false
-        profileViewController.logout.isHidden = true
-        profileViewController.line.isHidden = true
-        profileViewController.deleteID.isHidden = true
-
-         // 전체 화면으로 모달을 표시하려면 이 줄을 추가하세요.
-        self.present(profileViewController, animated: true, completion: nil)
+        PresentToProfileVC.presentToProfileVC(from: self, with: profile)
     }
-    
-    func getUserImage(referencePath: String?, imageSize: ImageSize, completion: @escaping(UIImage?) -> Void) {
-        guard let referencePath else { return }
-        let imageRefPath = storage.child(referencePath).child(imageSize.rawValue).fullPath
-        self.urlCache.downloadURL(storagePath: imageRefPath) { result in
-            switch result {
-            case .success(let image):
-                completion(image)
-            case .failure(let error):
-                print(error.localizedDescription)
-            }
-        }
-    }
-    
 }
